netiquette.yml: slightly change the "chaining" block format
In "chaining", we combine blocks defined in the "block" section to tell Dave
how these blocks may occur in a valid email.

Using "<foo>" (same syntax as subrule calls in "grammar" values) to mean "the
'foo' block" may incorrectly lead users to think they can use all shiny
features from Perl 5 regexes, which they don't.

So we need another syntax. Not using any delimiter could be less readable.
"{" and "}" are used for quantification, and "(" and ")" for grouping, so
we're left with "[" and "]", which I like here.